+++ src/internet-apps/helper/radvd-helper.cc
@@ -27,7 +27,8 @@ RadvdHelper::RadvdHelper()
 void
 RadvdHelper::AddAnnouncedPrefix(uint32_t interface,
                                 const Ipv6Address& prefix,
-                                uint32_t prefixLength)
+                                uint32_t prefixLength,
+                                bool slaac)
 {
     NS_LOG_FUNCTION(this << int(interface) << prefix << int(prefixLength));
     if (prefixLength != 64)
@@ -59,6 +60,7 @@ RadvdHelper::AddAnnouncedPrefix(uint32_t interface,
     if (!prefixFound)
     {
         Ptr<RadvdPrefix> routerPrefix = Create<RadvdPrefix>(prefix, prefixLength);
+        routerPrefix->SetAutonomousFlag(slaac);
         m_radvdInterfaces[interface]->AddPrefix(routerPrefix);
     }
 }