+++ src/internet-apps/helper/radvd-helper.h
@@ -35,8 +35,12 @@ class RadvdHelper : public ApplicationHelper
      * @param interface outgoing interface
      * @param prefix announced IPv6 prefix
      * @param prefixLength announced IPv6 prefix length
+     * @param slaac available for autoconfiguration
      */
-    void AddAnnouncedPrefix(uint32_t interface, const Ipv6Address& prefix, uint32_t prefixLength);
+    void AddAnnouncedPrefix(uint32_t interface,
+                            const Ipv6Address& prefix,
+                            uint32_t prefixLength,
+                            bool slaac = true);
 
     /**
      * @brief Enable the router as default router for the interface.